:sparkles: new Dispose{} block! :sparkles: Add some new constructors for ScriptBlockAst
:wrench: Invoke Dispose{} with proper scope context
:heart: Thank you @seeminglyscience!

:wrench: Prevent Dispose{} from executing more than once.
This is done by ensuring that the dispose operations complete even if
the code in Dispose{} throws errors.

:wrench: Allow Dispose{} to throw errors correctly.
Previously Dispose{} was permitting only non-terminating errors through.
This adds handling for terminating errors thrown from Dispose{} and
ensures that they are handled similarly to those thrown from other named
blocks.

:sparkles: Dispose commands after EndProcessing()
This ensures that commands are disposed once they're done with, without
waiting for the whole pipeline to complete.
This should enable some slightly easier ways to work with cmdlets that
need to work with the same files or other resources that would otherwise
be retained by cmdlets earlier in the pipeline.

:sparkles: Add case for visiting Dispose{} and refactor visit code.

:recycle: Add new ShouldContinueAfterVisit function that allows us to
avoid unnecessarily repeating the check whether we should continue.
As soon as the visitor reaches a point where the next action is not
Continue, the search for other named blocks ceases.

:sparkles: Permit use of dispose in functions

:sparkles: Allow breakpoints in Dispose{}