:bug: Fix string param conversion for bigints
When passing a bigint literal, e.g., `7n` as a bare string to a
parameter of type [string], the parameter binder was incorrectly using
the direct ToString() conversion unlike other numeric literals, which
store the original TokenText in order to properly retrieve the original
string that was entered at the command line.

Fix is to ensure that we're capturing and storing the TokenText value
when the value is of type BigInteger (since this type's TypeCode is
simply `object` and is not read as being a numeric type by those APIs.

We also need to add BigInteger to the array of numeric types stored in
LanguagePrimitives in order for the conversion methods to recognise
that BigInteger literals need to be converted to string more carefully.